Clear up codes
