+++ rayvision_api/operators/__init__.py
@@ -1,16 +1,15 @@
 """The operations of the rayvision_api."""
 
 from rayvision_api.operators.env import RenderEnvOperator
-from rayvision_api.operators.query import QueryOperator
-from rayvision_api.operators.tag import TagOperator
-from rayvision_api.operators.task import TaskOperator
+from rayvision_api.operators.project import ProjectOperator
+from rayvision_api.operators.job import JobOperator
 from rayvision_api.operators.user import UserOperator
 
 # All public api.
 __all__ = (
     'RenderEnvOperator',
     'QueryOperator',
-    'TagOperator',
-    'TaskOperator',
+    'ProjectOperator',
+    'JobOperator',
     'UserOperator'
 )
