+++ rayvision_api/operators/job.py
@@ -1,80 +1,216 @@
-# -*- coding: utf-8 -*-
-"""API query operation."""
-import sys
-
-from rayvision_api import constants
+"""Interface to operate on the task.
 
 
-class QueryOperator(object):
-    """API query operation."""
+
+"""
+
+from rayvision_api.exception import RayvisionError
+
+
+class JobOperator(object):
+    """API task related operations."""
+
+    TASK_PARAM = "taskIds"
 
     def __init__(self, connect):
         """Initialize instance.
 
         Args:
-            connect (rayvision_api.api.connect.Connect): The connect instance.
+            connect (rayvision_api.connect.Connect): The connect instance.
 
         """
         self._connect = connect
+        self._has_submit = False
+        self._task_id = None
 
-    def platforms(self):
-        """Get platforms.
-
-        Returns:
-            list: Platforms info.
-                e.g.:
-                     [
-                         {
-                             "platform": 2,
-                             "name": "query_platform_w2"
-                         },
-                     ]
-
-        """
-        zone = 1 if "renderbus" in self._connect.domain.lower() else 2
-        return self._connect.post(self._connect.url.queryPlatforms,
-                                  {'zone': zone})
-
-    def error_detail(self, code, language='0'):
-        r"""Get analysis error code.
+    def _create_task(self,
+                     count=1,
+                     task_user_level=50,
+                     out_user_id=None,
+                     labels=None):
+        """Create a task ID.
 
         Args:
-            code (int): Required value, error code.
-                e.g.:
-                    10010.
-                    15000.
-            language (str, optional): Not required, language,
-                0: Chinese (default) 1: English.
+            count (int, optional): The quantity of task ID.
+            task_user_level (int): Set the user's task level to either 50 or
+                60, default is 50.
+            out_user_id (int, optional): Non-required, external user ID, used
+                to distinguish users accessing third parties.
+            labels (list or tuple, optional): Custom task labels.
 
         Returns:
-            list: Detailed list of error messages.
+            dict: The information of the task.
                 e.g.:
-                     [
-                         {
-                            "id": 5,
-                            "code": "15000",
-                            "type": 1,
-                            "languageFlag": 0,
-                            "desDescriptionCn": "启动 3ds max 卡住或者失败",
-                            "desSolutionCn": "1.检查启用对应版本的 3ds max
-                                是否有特殊弹窗，有的话手动关闭；\n2.检查操作系
-                                统是否设置了高级别的权限",
-                            "solutionPath": "
-                                http://note.youdao.com/noteshare?id=d8f1ea0c46dfb524af798f6b1d31cf6f",
-                            "isRepair": 0,
-                            "isDelete": 1,
-                            "isOpen": 1,
-                            "lastModifyAdmin": "",
-                            "updateTime": 1534387709000
-                         },
-                     ]
+                    {
+                        "taskIdList": [1658434],
+                        "aliasTaskIdList": [2W1658434],
+                        "userId": 100093088
+                    }
 
         """
         data = {
-            'code': code,
-            'language': language
+            'count': count,
+            'taskUserLevel': task_user_level
         }
-        return self._connect.post(self._connect.url.queryErrorDetail, data)
+        if out_user_id:
+            data['outUserId'] = out_user_id
+        if labels:
+            data['labels'] = labels
+        return self._connect.post(self._connect.url.createTask, data)
+
+    def _generate_task_id(self):
+        """Get task id.
+
+        Example::
+
+            task_id_info = {
+                    "taskIdList": [1658434],
+                    "aliasTaskIdList": [2W1658434],
+                    "userId": 100093088
+                }
+
+        Returns:
+            int: The ID number of the task.
+
+        """
+        if not self._has_submit and self._task_id:
+            return self._task_id
+        task_id_info = self._create_task(count=1, out_user_id=None)
+        task_id_list = task_id_info.get("taskIdList")
+        if not task_id_list:
+            raise RayvisionError(1000000, 'Failed to create task number!')
+        self._task_id = task_id_list[0]
+        self._has_submit = False
+        return self._task_id
+
+    @property
+    def task_id(self):
+        """int: The ID number of the render task.
+
+        Notes:
+            As long as we do not initialize the class again or submit the task
+            successfully, we can always continue to get the task id from the
+            class instance.
+
+        """
+        return self._generate_task_id()
+
+    def submit_task(self,
+                    info,
+                    asset_lsolation_model=None,
+                    out_user_id=None,
+                    only_id=False):
+        """Submit a task to rayvision render farm.
+
+        Args:
+            task_id (int): Submit task ID.
+            asset_lsolation_model (str): Asset isolation type, Optional value,
+                default is null, optional value:'TASK_ID_MODEL' or 'OUT_USER_MODEL'.
+            out_user_id (str): The asset isolates the user ID, Optional value,
+                when asset_lsolation_model='OUT_USER_MODEL' ,'out_user_id'
+                cant be empty.
+
+        """
+        data = {
+            "taskId": self.task_id
+        }
+        if asset_lsolation_model:
+            data["assetIsolationModel"] = asset_lsolation_model
+        if out_user_id:
+            data["outUserId"] = out_user_id.strip()
+        data.update(info)
+        task_info = self._connect.post(self._connect.url.submitTask, data)
+        if only_id:
+            return self.task_id
+        self._has_submit = True
+        return task_info
+
+    def stop_task(self, task_param_list):
+        """Stop the task.
+
+        Args:
+            task_param_list (list): Task ID list.
+
+        """
+        return self._connect.post(self._connect.url.stopTask,
+                                  {self.TASK_PARAM: task_param_list})
+
+    def start_task(self, task_param_list):
+        """Start task.
+
+        Args:
+            task_param_list (list): Task ID list.
+
+        """
+        return self._connect.post(self._connect.url.startTask,
+                                  {self.TASK_PARAM: task_param_list})
+
+    def abort_task(self, task_param_list):
+        """Give up the task.
+
+        Args:
+            task_param_list (list): Task ID list.
+
+        """
+        return self._connect.post(self._connect.url.abortTask,
+                                  {self.TASK_PARAM: task_param_list})
+
+    def delete_task(self, task_param_list):
+        """Delete task.
+
+        Args:
+            task_param_list (list): Task ID list.
+
+        """
+        return self._connect.post(self._connect.url.deleteTask,
+                                  {self.TASK_PARAM: task_param_list})
+
+    def update_priority(self, task_id, priority):
+        """Update the render priority for the task by given task id.
+
+        Args:
+            task_id (int): The ID number of the render task.
+            priority (int): The priority for the current render task.
+
+        """
+        data = {
+            'taskId': task_id,
+            'taskUserLevel': priority,
+        }
+        return self._connect.post(self._connect.url.updateTaskUserLevel, data)
+
+    def set_task_overtime_top(self, task_id_list, overtime):
+        """Set the task timeout stop time.
+
+        Args:
+            task_id_list (list of int): Task list.
+            overtime (int or float): Timeout time, unit: second.
+            Example:
+                {
+                    "taskIds":[485],
+                    "overTime":1800
+                }
+        """
+        data = {
+            'taskIds': task_id_list,
+            'overTime': overtime
+        }
+        return self._connect.post(self._connect.url.setOverTimeStop, data)
+
+    def full_speed(self, task_id_list):
+        """Full to render.
+
+        Args:
+            task_id_list (list of int): Task list.
+            Example:
+                {
+                    "taskIds":[485],
+                }
+        """
+        data = {
+            'taskIds': task_id_list,
+        }
+        return self._connect.post(self._connect.url, data)
 
     def get_task_list(self, page_num=1, page_size=2, status_list=None,
                       search_keyword=None,
@@ -86,13 +222,14 @@ class QueryOperator(object):
         Args:
             page_num (int): Required value, current page.
             page_size (int): Required value, numbers displayed per page.
-            status_list (list<int>): status code list，query the status of the task in the list.
+            status_list (list of int): status code list, query the status of
+                the task in the list.
             search_keyword (string): Optional, scenario name or job ID.
             start_time (string): Optional, search limit for start time.
             end_time (string): Optional, search limit for end time.
 
         Returns:
-            dict: Task info, please see the documentation for details.
+            dict: Task profile, please see the documentation for details.
                 e.g.:
                     {
                         "pageCount": 32,
@@ -101,7 +238,7 @@ class QueryOperator(object):
                         "size": 1,
                         "items": [
                             {
-                                "sceneName": "衣帽间.max",
+                                "sceneName": "",
                                 "id": 18278,
                                 "taskAlias": "P18278",
                                 "taskStatus": 0,
@@ -144,7 +281,7 @@ class QueryOperator(object):
                 rendering, optional.
 
         Returns:
-            dict: Frames info list, please see the documentation for details.
+            dict: Frames profile list, please see the documentation for details.
                 e.g.:
                     {
                         "pageCount": 9,
@@ -181,7 +318,7 @@ class QueryOperator(object):
         """Get the overview of task rendering frame.
 
         Returns:
-            dict: Frames status info.
+            dict: Frames status profile.
                 e.g.:
                     {
                         "executingFramesTotal": 1,
@@ -205,7 +342,7 @@ class QueryOperator(object):
         data = {
             'taskIds': task_param_list
         }
-        return self._connect.post(URL.restartFailedFrames, data)
+        return self._connect.post(self._connect.url.restartFailedFrames, data)
 
     def restart_frame(self, task_id, select_all, ids_list=None):
         """Re-submit the specified frame.
@@ -294,113 +431,43 @@ class QueryOperator(object):
         }
         return self._connect.post(self._connect.queryTaskInfo, data)
 
-    def supported_software(self):
-        """Get supported rendering software.
-
-        Returns:
-            dict: Software info.
-                e.g.:
-                    {
-                        "isAutoCommit": 2,
-                        "renderInfoList": [
-                            {
-                                "cgId": 2000,
-                                "cgName": "Maya",
-                                "cgType": "ma;mb",
-                                "iconPath": "/img/softimage/maya.png",
-                                "isNeedProjectPath": 3,
-                                "isNeedAnalyse": 1,
-                                "isSupportLinux": 1
-                            }
-                        ],
-                        "defaultCgId": 2001
-                    }
-
-        """
-        return self._connect.post(self._connect.url.querySupportedSoftware,
-                                  validator=False)
-
-    def supported_plugin(self, name):
-        """Get supported rendering software plugins.
+    def error_detail(self, code, language='0'):
+        """Get analysis error code.
 
         Args:
-            name (str): The name of the DCC.
+            code (int): Required value, error code.
                 e.g.:
-                    maya,
-                    houdini
+                    10010.
+                    15000.
+            language (str, optional): Not required, language,
+                0: Chinese (default) 1: English.
 
         Returns:
-            dict: Plugin info.
+            list of dict: Detailed list of error messages.
                 e.g.:
-                    {
-                        "cgPlugin": [
-                            {
-                                "cvId": 19,
-                                "pluginName": "zblur",
-                                "pluginVersions": [
-                                    {
-                                        "pluginId": 1652,
-                                        "pluginName": "zblur",
-                                        "pluginVersion": "zblur 2.02.019"
-                                    }
-                                ]
-                            },
-                        ],
-                        "cgVersion": [
-                            {
-                                "id": 23,
-                                "cgId": 2005,
-                                "cgName": "CINEMA 4D",
-                                "cgVersion": "R19"
-                            }
-                        ]
-                    }
+                     [
+                         {
+                            "id": 5,
+                            "code": "15000",
+                            "type": 1,
+                            "languageFlag": 0,
+                            "desDescriptionCn": "",
+                            "desSolutionCn": ""
+                            "solutionPath": "",
+                            "isRepair": 0,
+                            "isDelete": 1,
+                            "isOpen": 1,
+                            "lastModifyAdmin": "",
+                            "updateTime": 1534387709000
+                         },
+                     ]
 
         """
-        cg_id = constants.DCC_ID_MAPPINGS[name.strip().lower()]
-        platform = "windows" if sys.platform.startswith("win") else "linux"
-        data = {'cgId': cg_id, 'osName': platform}
-        return self._connect.post(self._connect.url.querySupportedPlugin, data)
-
-    def get_transfer_server_msg(self):
-        """Get the user rendering environment configuration.
-
-        Returns:
-            dict: Connect raysync information.
-                Example:
-                    {
-                        'raysyncTransfer': {
-                            'port': 2542,
-                            'proxyIp': 'render.raysync.cn',
-                            'proxyPort': 32011,
-                            'serverIp': '127.0.0.1',
-                            'serverPort': 2121,
-                            'sslPort': 2543
-                        }
-                    }
-
-        """
-        zone = 1 if "renderbus" not in self._connect.domain else 2
         data = {
-            "zone": zone
+            'code': code,
+            'language': language
         }
-        return self._connect.post(self._connect.url.getTransferServerMsg, data)
-
-    def get_raysync_user_key(self):
-        """Get the user rendering environment configuration.
-
-        Returns:
-            dict: User login raysync information.
-                Example:
-                    {
-                        'raySyncUserKey': '8ccb94d67c1e4c17fd0691c02ab7f753cea64e3d',
-                        'userName': 'test',
-                        'platform': 2,
-                    }
-
-        """
-        return self._connect.post(self._connect.url.getRaySyncUserKey,
-                                  validator=False)
+        return self._connect.post(self._connect.url.queryErrorDetail, data)
 
     def get_task_processing_img(self, task_id, frame_type=None):
         """Get the task progress diagram,currently only Max software is supported.
