+++ requirements.txt
@@ -1,6 +1,6 @@
-attrs==19.3.0
-cattrs==1.0.0
+enum34==1.1.10
 tenacity==6.0.0
 future==0.18.2
 requests==2.23.0
+cerberus==1.3
 rayvision_log>=0.3.3
